Removed unused extension methods and moved to server proj
Created new class for extension methods, as I couldn't find any others being used in the project.
